+++ src/app/signup/web3AuthUtils.ts
